Move enums from endpoint to connection
